updated nav and show player position

+++ rs_player.h
@@ -3,7 +3,7 @@
 
 #include "rs_types.h"
 
-rs_player* rs_make_player(u32 map_x, u32 map_y);
+rs_player* rs_make_player(float map_x, float map_y);
 
 void rs_free_player(rs_player* p);
